+++ tests/testthat/test-extract.R
@@ -22,15 +22,12 @@ car_set_1 <-
 # ------------------------------------------------------------------------------
 
 test_that("extracts", {
-   expect_snapshot(error = TRUE,
-      extract_fit_engine(car_set_1, id = "reg_lm")
-   )
-   expect_snapshot(error = TRUE,
-      extract_fit_parsnip(car_set_1, id = "reg_lm")
-  )
-   expect_snapshot(error = TRUE,
-    extract_mold(car_set_1, id = "reg_lm")
-  )
+  # workflows specific errors, so we don't capture their messages
+  expect_error(extract_fit_engine(car_set_1, id = "reg_lm"))
+  expect_error(extract_fit_parsnip(car_set_1, id = "reg_lm"))
+  expect_error(extract_mold(car_set_1, id = "reg_lm"))
+  expect_error(extract_recipe(car_set_1, id = "reg_lm"))
+
   expect_s3_class(
     extract_preprocessor(car_set_1, id = "reg_lm"),
     "recipe"
@@ -42,9 +39,6 @@ test_that("extracts", {
   expect_s3_class(
     extract_workflow(car_set_1, id = "reg_lm"),
     "workflow"
-  )
-  expect_snapshot(error = TRUE,
-    extract_recipe(car_set_1, id = "reg_lm")
   )
   expect_s3_class(
     extract_recipe(car_set_1, id = "reg_lm", estimated = FALSE),
@@ -61,13 +55,13 @@ test_that("extracts", {
     car_set_1$result[[1]]
   )
 
-  expect_snapshot(error = TRUE,
-    car_set_1 %>% extract_workflow_set_result("Gideon Nav")
-  )
+  expect_snapshot(error = TRUE, {
+     car_set_1 %>% extract_workflow_set_result("Gideon Nav")
+  })
 
-  expect_snapshot(error = TRUE,
-    car_set_1 %>% extract_workflow("Coronabeth Tridentarius")
-  )
+  expect_snapshot(error = TRUE, {
+     car_set_1 %>% extract_workflow("Coronabeth Tridentarius")
+  })
 })