+++ NEWS.md
@@ -1,5 +1,7 @@
 # workflowsets (development version)
 
+* Updated tests related to changes in workflows 0.2.5 (#75).
+
 * `as_workflow_set()` can now take a mixture of workflows or `tune_results` objects. 
 
 * New `extract_parameter_set_dials()` and `extract_parameter_dials()` methods to